Create send_script.php
When the user clicks the send button on the contact form, the values in the form are submitted to the send_script.php file. This file grabs the values from the form and forwards them to the email address specified. The email can be sent in HTML format as above and CSS can be used to apply some styles to the email. 

Note:
Sending email using the mail() function only works when there is internet connection. Until your application is hosted on a server on the internet before it can send an email.